+++ packages/office-ui-fabric-react/src/utilities/buttonGrid/buttonGrid.styles.ts
@@ -1,0 +1,13 @@
+import { IButtonGridStyleProps, IButtonGridStyles } from './buttonGrid.types';
+
+export const getStyles = (props: IButtonGridStyleProps): IButtonGridStyles => {
+  return {
+    root: {
+      padding: 2,
+      outline: 'none',
+    },
+    tableCell: {
+      padding: 0,
+    },
+  };
+};
