+++ packages/office-ui-fabric-react/src/utilities/buttonGrid/index.ts
@@ -1,0 +1,4 @@
+export * from './buttonGrid';
+export * from './buttonGrid.types';
+export * from './buttonGridCell';
+export * from './buttonGridCell.types';
